chore(deployment): use dragonfly configmap for image
Previously the image was unspecified, leaving it up to the operator to
control the version. We'd like tighter control over this, separate from
our update cadence of the operator itself. This moves to using a value
from the `dragonfly` configmap which can be shared (or overridden) by
individual instances, for vector or otherwise.

Ref: LOG-21944