feat(sources): Configuration to control Pulsar nacking
When events fail to deliver to a sink, the Pulsar messages were being
nacked. This causes redelivery from the broker causing the pipeline to
process the same events over and over, which may break aggregations.
This commit adds a configuration option to turn this behavior off (by
default) so that pipelines only see the message once and rely on sink
retry mechanisms for it to reach the destination.

Ref: LOG-22106