fix(sinks): Rename opentelemetry to mezmo_opentelemetry
In Vector 0.43.0, upstream has added their own version of an
opentelemetry sink. This will conflict with a sink of the same name
which we added long ago.  To avoid future merge conflicts, try to rename
our version to `mezmo_opentelemetry` and we will disable the build for
their future `opentelemetry` sink to avoid conflicts with the `type`
name we already use in `pipeline-service`.

Ref: LOG-22131