fix(datadog): ensure metrics version in host is valid [skip lint]
Our feature builds get a (valid) semver like `1.2.3+feature-123-1`. The
datadog metrics sink includes the version as a wildcard subdomain
component, but does not anticipate more exotic versions such as our
feature builds. The `+` makes the host invalid and this does not pop an
error until the time of the request.

Ref: LOG-22128