feat(transforms): add persistence to mezmo_reduce
Adds support for persistence in RocksDB for the `mezmo_reduce`
transform. This allows the component state to survive restarts,
and also provides an ack boundary such that finalizers are dropped
preemptively which avoids building lag over long reduce periods.

Most of the changes here are to support serializing and deserializing
the state to/from RocksDB.

Ref: LOG-22359