feat(sources): `pulsar` source should auto-detect new partitions
When using partitioned topics, an administrator may choose to add more
partitions to a topic to help balance load. When the pulsar consumer is
built, it gathers a static list of topics which doesn't change when
using `.with_topics()`. Therefore, new partitions would not be picked up
without a restart of the pipeline or vector. This commit adds using the
functionality for topic regexes to pick up new partitions if they are
added.

Ref: LOG-22427