feat(sources): Pulsar source should accept connection retry options
The pulsar client has connection retry options which should be exposed
by the vector source. This commit also uses a default of `u32::MAX` for
`max_retries` so that Pulsar never drops the connection due to failure.
Otherwise, Vector will stop processing, even upon reconnection. Such
cases are when the `pulsar-proxy` is down, which can take several
minutes to stabilize. Pulsar's default of `12` retries is not enough to
tolerate connection instability.

Ref: LOG-22575