fix(pulsar): Upgrade pulsar 6.5.0 and patch tokio-util
We want to upgrade to pulsar@6.5.0, but it contains an updated
dependency for tokio-util which breaks vector. Vector has maintained
a fork of `tokio` and patched it, so this same patch was applied to the
version that pulsar 6.5.0 uses, which is `tokio-util@0.7.17`.

We are testing this version of pulsar as a potential fix for the
`SlowDown` error which causes partitions to no longer be consumed. We
need consumers to be pulling from all available partitions.

Ref: LOG-22655