fix(sources): Reorganize pulsar tokio::select block
This pulsar consumer seemed to error when the brokers were rolled, such
that vector would stop consuming from the topic. Re-organizing the
`tokio::select` block to get the full result of the `consumer.next()`
call allow the `None` arm to be checked properly. Changing it this way
appears to have allowed the pulsar client to auto-heal through broker
and proxy crashes.

Ref: LOG-22768